+++ interpreter/main.py
@@ -3,44 +3,7 @@ from antlr4 import *
 from cash.CASHParser import CASHParser
 from cash.CASHLexer import CASHLexer
 from cash.CASHVisitor import CASHVisitor
-
-
-class Task: 
-    def __init__(self, name: str, param_names: list, body):
-        self.name = name
-        self.param_names = param_names
-        self.body = body
-
-    def execute(self, visitor: 'InterpreterVisitor', param_values): 
-
-    
-        for name, value in zip(self.param_names, param_values):
-            visitor.symbol_table.add_var(name, value)
-
-        for stmt in self.body:
-            visitor.visit(stmt)
-
-
-class SymbolTable:
-
-    def __init__(self):
-        self.storage = {}
-        self.tasks = {}
-
-    def add_var(self, name: str, value: float):
-        self.storage[name] = value
-
-    def get_var(self, name: str):
-        if name not in self.storage:
-            raise KeyError(f"Variable {name} not found!!")
-        return self.storage[name]
-    
-
-    def add_task(self, task: Task):
-        self.tasks[task.name] = task 
-
-    def get_task(self, name: str): 
-        return self.tasks[name]
+from symbol_table import SymbolTable, Task
 
 class InterpreterVisitor(CASHVisitor): 
     def __init__(self, symbol_table: SymbolTable):
@@ -166,6 +129,14 @@ class InterpreterVisitor(CASHVisitor):
 
 def main():
     input_stream = FileStream("./example_code/helloWorld.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/calc.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/fibonacci.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/func.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/ifThenElse.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/int_float.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/promptUser.csh", encoding="utf-8")
+    # input_stream = FileStream("./example_code/while.csh", encoding="utf-8")
+
     lexer = CASHLexer(input_stream)
     token_stream = CommonTokenStream(lexer)
     parser = CASHParser(token_stream)
